Bandcamp Friday Website JSON Parsing Fix (#26)
We were parsing JSON weirdly since we didn't want to add another
dependency or something. Not really sure why I did it that way in the
first place...

Now we use the `json` library to parse it so we know we are getting the
`"display"` property on the `data-fundraisers` attribute.

Additionally, it looks like we are getting Bandcamp Friday notifications
1 day early. Our pervious implementation required setting the
notification back 1 day, but apparently this new implementation does not
require that. Probably a time-zone issue that I could make sense of but
instead we will just offset (or remove offset) until it works.