Bandcamp Friday: Continue Updates During Failure; Add Debug Logs (#28)
The github action runner appears to be getting blocked from accessing
the "Is it Bandcamp Friday" website. Locally, we can still access the
site and it works as expected, finding the next date (which is actually
in the past, implying that there are no future Bandcamp Fridays planned.

For now, make the code "safe" against this kind of issue and print out
any information we can use to fix this.