Don't Remove Trackable Items (#33)
We have had issues in the past with Bandcamp items re-appearing in the
RSS feed. This is due to the items in the RSS file getting removed one
day and re-added the next day. The RSS feed does not support notifying
when items are removed from wishlists, releases, etc., so we shouldn't
be tracking that in the SSFs either.
- Following items still track deletes as we use that file to determine
which releases to check for
- This means that wishlist, release, and collection items no longer
track deletes, even when you unfollow an artist

We sometimes get "0 bytes" from querying a link from Bandcamp. This is
likely a rate limit. To be respectful of Bancamp's APIs/website, we
added a retry that waits up to 60 seconds before querying bandcamp
again.
- This can also occur if we ping the wrong URL for whatever reason

These changes also account for artists with only 1 tralbum or artists
with no tralbums. This is mainly so that you can follow an artist with 0
or 1 tralbum, and when they release their first or second tralbum, only
that new one will show up. Previously an artist with only 1 tralbum that
released their second one would result in 2 notifications, and an artist
with 0 tralbums that released 1 would not appear at all.
- Note that since this is a change in behavior, we have decided to skip
notifying these artists during the initial deployment phase